requirements.txt
torch>=1.8 # Or higher, check CUDA compatibility
torchvision
opencv-python-headless # Use headless unless you need GUI features
pandas
numpy
scikit-learn
matplotlib
tqdm
kagglehub # Optional, for automatic download
Pillow # PIL fork, often installed with torch/torchvision